Optimize package install by testing if package installed before attempting to install it.

+++ vagrant/solarnet-dev/bootstrap.sh
@@ -79,14 +79,18 @@ sudo apt-get update
 echo -e '\nUpgrading outdated packages...'
 sudo DEBIAN_FRONTEND=noninteractive apt-get upgrade -qy
 
-echo -e '\nInstalling language-pack...'
-sudo DEBIAN_FRONTEND=noninteractive apt-get install -qy language-pack-en
+if ! dpkg -s language-pack-en >/dev/null 2>/dev/null; then
+	echo -e '\nInstalling language-pack...'
+	sudo DEBIAN_FRONTEND=noninteractive apt-get install -qy language-pack-en
+fi
 
-echo -e '\nInstalling git...'
-sudo DEBIAN_FRONTEND=noninteractive apt-get install -qy git git-lfs
-if ! grep -q lfs ~/.gitconfig >/dev/null 2>/dev/null; then
-	echo -e '\nInitializing git LFS...'
-	git lfs install --skip-repo
+if ! dpkg -s git-flow >/dev/null 2>/dev/null; then
+	echo -e '\nInstalling git...'
+	sudo DEBIAN_FRONTEND=noninteractive apt-get install -qy git git-lfs git-flow
+	if ! grep -q lfs ~/.gitconfig >/dev/null 2>/dev/null; then
+		echo -e '\nInitializing git LFS...'
+		git lfs install --skip-repo
+	fi
 fi
 
 if [ -x /usr/share/postgresql-common/pgdg/apt.postgresql.org.sh ]; then
@@ -138,29 +142,36 @@ if [ -n "$DESKTOP_PACKAGES" ]; then
 fi
 
 for v in $JAVAVER; do
-	echo -e "\nInstalling Java $v..."
 	javaPkg=openjdk-$v-jdk
 	if [ -z "$DESKTOP_PACKAGES" ]; then
 		javaPkg="${javaPkg}-headless"
 	fi
-	sudo DEBIAN_FRONTEND=noninteractive apt install -qy $javaPkg
+	if ! dpkg -s $javaPkg >/dev/null 2>/dev/null; then
+		echo -e "\nInstalling Java $v..."
+		sudo DEBIAN_FRONTEND=noninteractive apt install -qy $javaPkg
+	fi
 done
   
-echo -e "\nInstalling supporting utilities..."
-sudo DEBIAN_FRONTEND=noninteractive apt install -qy gnupg apt-transport-https lsb-release wget \
-  git git-flow
+if ! dpkg -s git-flow >/dev/null 2>/dev/null; then
+	echo -e "\nInstalling supporting utilities..."
+	sudo DEBIAN_FRONTEND=noninteractive apt install -qy gnupg apt-transport-https lsb-release wget
+fi
   
-echo -e "\nInstalling Postgres $PGVER..."
-sudo DEBIAN_FRONTEND=noninteractive apt install -qy postgresql-$PGVER postgresql-contrib-$PGVER \
-  postgresql-common
+if ! dpkg -s postgresql-$PGVER >/dev/null 2>/dev/null; then
+	echo -e "\nInstalling Postgres $PGVER..."
+	sudo DEBIAN_FRONTEND=noninteractive apt install -qy postgresql-$PGVER postgresql-contrib-$PGVER \
+	  postgresql-common
+fi
 
-echo -e '\nInstalling Postgres extensions...'
-sudo apt install -qy timescaledb-2-postgresql-$PGVER postgresql-$PGVER-aggs-for-vecs
+if ! dpkg -s timescaledb-2-postgresql-$PGVER >/dev/null 2>/dev/null; then
+	echo -e '\nInstalling Postgres extensions...'
+	sudo apt install -qy timescaledb-2-postgresql-$PGVER postgresql-$PGVER-aggs-for-vecs
+fi
 
 echo -e '\nCleaning up unused packages...'
 sudo apt autoremove -qy
 
-if ! grep -q 'jit = on' /etc/postgresql/$PGVER/main/postgresql.conf 2>/dev/null; then
+if ! grep -q 'jit = on' /etc/postgresql/$PGVER/main/postgresql.conf >/dev/null 2>/dev/null; then
 	echo -e '\nDisabling JIT in Postgres...'
 	sudo sed -i -e 's/^#*jit = .*/jit = off/' /etc/postgresql/$PGVER/main/postgresql.conf
 	sudo service postgresql restart
@@ -316,7 +327,7 @@ fi
 sudo systemctl enable vernemq
 
 # Install Mosquitto client
-if dpkg -s mosquitto-clients 2>/dev/null; then
+if dpkg -s mosquitto-clients >/dev/null 2>/dev/null; then
 	echo -e '\nMosquitto MQTT client already installed.'
 else
 	echo -e '\nInstalling Mosquitto MQTT client...'
